# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ATANDA Command Centre + HARNESS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /harness/f1:
    post:
      operationId: harnessF1
      tags: [harness]
      summary: ATLAS Prompt Diagnostic
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF1Input" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PromptDiagnostic" }
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /harness/f2:
    post:
      operationId: harnessF2
      tags: [harness]
      summary: Atomic Prompt Builder certification
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF2Input" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AtomicPrompt" }
  /harness/f3:
    post:
      operationId: harnessF3Stream
      tags: [harness]
      summary: CELL MA Builder (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF3Input" }
      responses:
        "200":
          description: text/event-stream of MA cultivation events
          content:
            text/event-stream:
              schema: { type: string }
  /harness/f4:
    post:
      operationId: harnessF4
      tags: [harness]
      summary: Micro PDD converter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF4Input" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MicroPdd" }
  /harness/f5:
    post:
      operationId: harnessF5
      tags: [harness]
      summary: SPC Builder (FORGE Q&A turn)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF5Input" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/F5StepOutcome" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /harness/f5-finalize:
    post:
      operationId: harnessF5FinalizeStream
      tags: [harness]
      summary: SPC Builder — FORGE.COMMIT synthesis (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF5FinalizeInput" }
      responses:
        "200":
          description: text/event-stream of FORGE.COMMIT synthesis events
          content:
            text/event-stream:
              schema: { type: string }
        "403":
          $ref: "#/components/responses/Forbidden"
  /harness/f6:
    post:
      operationId: harnessF6
      tags: [harness]
      summary: ATLAS PDD Drafter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF6Input" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AtlasPdd" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /harness/f6-finalize:
    post:
      operationId: harnessF6DraftStream
      tags: [harness]
      summary: ATLAS PDD Drafter (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF6FinalizeInput" }
      responses:
        "200":
          description: text/event-stream of ATLAS PDD draft events
          content:
            text/event-stream:
              schema: { type: string }
        "403":
          $ref: "#/components/responses/Forbidden"
  /harness/f6-vdj:
    post:
      operationId: harnessF6Vdj
      tags: [harness]
      summary: VIBE DJ recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF6VdjInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VdjRecommendation" }
        "403":
          description: Tier insufficient (requires Practitioner+)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /harness/f05:
    post:
      operationId: harnessF05
      tags: [harness]
      summary: MATHMON Applicability Layer — intake report
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF05Input" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MathmonIntake" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /harness/map:
    post:
      operationId: harnessMapStream
      tags: [harness]
      summary: Mathematical Applicability Profile builder (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessMapInput" }
      responses:
        "200":
          description: text/event-stream of MAP generation events
          content:
            text/event-stream:
              schema: { type: string }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /harness/f7:
    post:
      operationId: harnessF7Stream
      tags: [harness]
      summary: SPARTAN Compressor (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF7Input" }
      responses:
        "200":
          description: text/event-stream of SPARTAN 7-step events
          content:
            text/event-stream:
              schema: { type: string }
        "403":
          $ref: "#/components/responses/Forbidden"
  /harness/f8:
    post:
      operationId: harnessF8
      tags: [harness]
      summary: Code DJ — scaffold a codebase from a certified MVP PDD (Architect tier)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF8Input" }
      responses:
        "200":
          description: Codebase bundle scaffolded from the source MVP PDD
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CodebaseBundle" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /harness/f9:
    post:
      operationId: harnessF9Mecha
      tags: [harness]
      summary: Run the deterministic F9 MECHA Machine Floor
      description: >
        Accepts an F8 CODEBASE_BUNDLE whose MVP PDD lineage is SPARTAN-certified.
        All seven phases and immutable external gate evidence must be supplied in
        order. Missing, non-passing, or unverifiable evidence produces a cited
        refusal; MECHA never invents or overrides a gate verdict.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF9Input" }
      responses:
        "200":
          description: Existing signed Machine Artifact returned for an idempotent repeat submission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/F9MachineArtifact" }
        "201":
          description: Signed Machine Artifact
          content:
            application/json:
              schema: { $ref: "#/components/schemas/F9MachineArtifact" }
        "400":
          description: Invalid phase envelope
        "409":
          description: Upstream F8/SPARTAN prerequisite failed
        "422":
          description: Cited MECHA refusal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/F9Refusal" }
  /harness/f9/runs:
    get:
      operationId: listHarnessF9Runs
      tags: [harness]
      summary: List owned F9 MECHA run records
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: F9 run state, evidence, and terminal outcome
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/F9RunRecord" }
  /harness/f8-hdj:
    post:
      operationId: harnessF8Hdj
      tags: [harness]
      summary: Host DJ — recommend a host + deployment journey from a certified MVP PDD (Architect tier)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessF8HdjInput" }
      responses:
        "200":
          description: Hosting plan produced from the certified MVP PDD (and optional codebase bundle)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HostingPlan" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Source MVP PDD is not SPARTAN-certified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /harness/atlas-crystallise:
    post:
      operationId: harnessAtlasCrystallise
      tags: [harness]
      summary: Crystallise a 4-Part ATLAS PDD into the typed ATLAS J JSON layer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/HarnessAtlasCrystalliseInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AtlasPddJson" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /harness/pfp:
    post:
      operationId: harnessPfp
      tags: [harness]
      summary: PDD Fidelity Protocol — cross-reference an MVP PDD against a codebase bundle (BUGMXT Layer 4 drift detection)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessPfpInput" }
      responses:
        "200":
          description: Drift report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PfpReport" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Source MVP PDD is not SPARTAN-certified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /harness/sessions/{id}/pdd-view:
    post:
      operationId: harnessPddView
      tags: [harness]
      summary: Render and persist an ATLAS 360 PLAN or SCAN view of the latest F6/F7 PDD
      description: >
        Read-only side-step over the latest owned ATLAS_PDD or MVP_PDD artifact.
        The persisted view never reruns F1-F8. Without format, manual and
        cartridge sessions default to PLAN and ingested sessions default to SCAN.
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [plan, scan] }
      responses:
        "200":
          description: Persisted ATLAS 360 PDD view
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Atlas360PddViewResponse" }
        "400":
          description: Invalid format
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "402":
          description: Monthly LLM cost cap reached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /harness/escalations/stream:
    get:
      operationId: harnessEscalationsStream
      tags: [harness]
      summary: SSE feed of escalations for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: text/event-stream
          content:
            text/event-stream:
              schema: { type: string }
...
  /harness/evolve:
    post:
      operationId: harnessEvolve
      tags: [harness]
      summary: DE-SPC — synthesise a Digitally Evolved SPC from N MA birth packages (requires ASPE badge)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HarnessEvolveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HarnessArtifact" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: atanda_staff
  responses:
...
    Forbidden:
      description: Forbidden (tier or role gate)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: ["string", "null"] }
...
    HarnessF8Input:
      type: object
      required: [sessionId, mvpPddArtifactId, platform]
      properties:
        sessionId: { type: string, format: uuid }
        mvpPddArtifactId: { type: string, format: uuid }
        platform: { $ref: "#/components/schemas/CodeDjPlatform" }
        notes:
          type: string
          description: Optional operator hints for the Code DJ (preferred libs, naming, etc.).
          maxLength: 2000
        acknowledgeDrift:
          type: boolean
          description: Override the PFP drift gate. Required when the most recent PFP report for the same MVP PDD has any critical findings.
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    HarnessF9Input:
      type: object
      required: [sessionId, sourceArtifactId, deviceClass, phases]
      properties:
        sessionId: { type: string, format: uuid }
        sourceArtifactId: { type: string, format: uuid }
        deviceClass: { type: string, minLength: 1, maxLength: 200 }
        artifactVersion: { type: string, pattern: '^[0-9]+\.[0-9]+\.[0-9]+$', default: "1.0.0" }
        phases:
          type: array
          minItems: 7
          maxItems: 7
          items: { $ref: "#/components/schemas/F9Phase" }
    F9Refusal:
      type: object
      required: [verdict, mecha_run_id, phase_halted, constraint_cited, invariant_cited, cause, required_to_proceed, gro_mode]
      properties:
        verdict: { type: string, enum: [REFUSED] }
        mecha_run_id: { type: string }
        phase_halted: { type: integer, minimum: 1, maximum: 7 }
        constraint_cited: { type: string }
        invariant_cited: { type: string }
        cause: { type: string }
        required_to_proceed: { type: string }
        gro_mode: { type: string, enum: [HUMAN_IN_LOOP, CONTAINMENT, KILLZONE] }
    F9MachineArtifact:
      type: object
      required: [machine_artifact_id, mecha_run_id, artifact_version, spk_id, mm_verdict, ucg_certificate, osiris_custody, reverification_due, payload_hash, artifact_signature, regulatory_conformity_asserted]
      properties:
        machine_artifact_id: { type: string }
        mecha_run_id: { type: string }
        artifact_version: { type: string }
        source_artifact_id: { type: string, format: uuid }
        spk_id: { type: string }
        mm_verdict: { type: string, enum: [MATH_VERIFIED] }
        ucg_certificate: { type: object, additionalProperties: true }
        savant_verdict: { type: string, enum: [FIT, CLUSTER, N/A] }
        osiris_custody: { type: boolean, enum: [true] }
        reverification_due: { type: string, format: date-time }
        payload_hash: { type: string, description: SHA-256 hash of canonical immutable payload }
        artifact_signature: { type: string, description: HMAC signature; signing key is never returned }
        regulatory_conformity_asserted: { type: boolean, enum: [false] }
        evidence: { type: object, additionalProperties: true }
    F9RunRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        mechaRunId: { type: string }
        sessionId: { type: string, format: uuid }
        sourceArtifactId: { type: string, format: uuid }
        evidenceHash: { type: string, description: SHA-256 hash of the canonical seven-phase evidence envelope }
        idempotencyKey: { type: string, description: Stable SHA-256 key derived from source artifact, artifact version, and evidence hash }
        status: { type: string, enum: [RUNNING, EMITTED, REFUSED] }
        phase: { type: integer }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/F9Phase" }
        refusal: { $ref: "#/components/schemas/F9Refusal" }
        payloadHash: { type: string }
        artifactSignature: { type: string }
        artifactContent: { $ref: "#/components/schemas/F9MachineArtifact" }
        osirisCustody: { type: boolean }
...
    HarnessF8HdjInput:
      type: object
      required: [sessionId, mvpPddArtifactId]
      properties:
        sessionId: { type: string, format: uuid }
        mvpPddArtifactId: { type: string, format: uuid }
        codebaseBundleArtifactId:
          type: string
          format: uuid
          description: Optional F8 codebase bundle to sharpen the hosting requirements profile.
        notes:
          type: string
          description: Optional operator hints for the Host DJ (preferred host, region/compliance constraints, budget ceiling).
          maxLength: 2000
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    HostingPlan:
      type: object
      required:
        [artifactId, hrp, weights, hse, primary, fallback, journey, sdf, jcse]
      properties:
        artifactId: { type: string, format: uuid }
        hrp: { $ref: "#/components/schemas/HostingRequirementsProfile" }
        weights: { $ref: "#/components/schemas/HseWeights" }
        hse:
          type: array
          description: Candidate hosts ranked desc by weightedTotal.
          items: { $ref: "#/components/schemas/HseRow" }
        primary: { $ref: "#/components/schemas/HostPick" }
        fallback: { $ref: "#/components/schemas/HostPick" }
        journey:
          type: object
          required: [tier, phases]
          properties:
            tier:
              {
                type: string,
                description: "Recommended deployment tier on the primary host.",
              }
            phases:
              type: array
              items: { $ref: "#/components/schemas/DeploymentJourneyPhase" }
        sdf: { $ref: "#/components/schemas/SelfDeployFactory" }
        jcse:
          {
            type: number,
            description: "Journey Confidence / Self-deploy Estimate 0..100, derived server-side from the winning weighted score.",
          }
        notes: { type: string }
    HarnessAtlasCrystalliseInput:
      type: object
      required: [sessionId, atlasPddArtifactId]
      properties:
        sessionId: { type: string, format: uuid }
        atlasPddArtifactId: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    AtlasPddJson:
      type: object
      required:
        [
          artifactId,
          schemaVersion,
          title,
          summary,
          prompts,
          stack,
          routes,
          deployTarget,
        ]
      properties:
        artifactId: { type: string, format: uuid }
        schemaVersion: { type: string, enum: ["atlas-pdd-v1"] }
        title: { type: string }
        summary: { type: string }
        prompts:
          type: array
          items: { $ref: "#/components/schemas/AtlasPddJsonPrompt" }
        stack:
          type: array
          items: { type: string }
        routes:
          type: array
          items: { type: string }
        deployTarget: { type: string }
...
    HarnessPfpInput:
      type: object
      required: [sessionId, mvpPddArtifactId, codebaseBundleArtifactId]
      properties:
        sessionId: { type: string, format: uuid }
        mvpPddArtifactId: { type: string, format: uuid }
        codebaseBundleArtifactId: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    CodebaseBundle:
      type: object
      required: [artifactId, platform, manifest, files, notes]
      properties:
        artifactId: { type: string, format: uuid }
        platform: { $ref: "#/components/schemas/CodeDjPlatform" }
        manifest: { $ref: "#/components/schemas/CodebaseManifest" }
        files:
          type: array
          items: { $ref: "#/components/schemas/CodebaseFile" }
        notes: { type: string }
...
    LlmProvider:
      type: string
      enum: [claude, openai, gemini]
      description: |
        LLM provider for HARNESS engine calls. Defaults to `claude`. Non-claude
        providers (`openai`, `gemini`) require PRACTITIONER tier or higher.
...
    PromptDiagnostic:
      type: object
      required: [jcse, certTier, pillars, atomicPrompt, strengths, gaps]
      properties:
        jcse: { $ref: "#/components/schemas/JcseBreakdown" }
        certTier: { $ref: "#/components/schemas/CertTier" }
        pillars:
          type: array
          items: { $ref: "#/components/schemas/PillarAssessment" }
        atomicPrompt: { $ref: "#/components/schemas/AtomicPromptTuple" }
        strengths:
          type: array
          items: { type: string }
        gaps:
          type: array
          items: { type: string }
        artifactId: { type: ["string", "null"], format: uuid }
    AtomicPrompt:
      type: object
      required: [tuple, jcse, certTier]
      properties:
        tuple: { $ref: "#/components/schemas/AtomicPromptTuple" }
        jcse: { $ref: "#/components/schemas/JcseBreakdown" }
        certTier: { $ref: "#/components/schemas/CertTier" }
        artifactId: { type: ["string", "null"], format: uuid }
...
    HarnessF1Input:
      type: object
      required: [sessionId, prompt]
      properties:
        sessionId: { type: string, format: uuid }
        prompt: { type: string, minLength: 10 }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    HarnessF2Input:
      type: object
      required: [sessionId, tuple]
      properties:
        sessionId: { type: string, format: uuid }
        tuple: { $ref: "#/components/schemas/AtomicPromptTuple" }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    HarnessF3Input:
      type: object
      required: [sessionId, atomicPrompt]
      properties:
        sessionId: { type: string, format: uuid }
        atomicPrompt: { $ref: "#/components/schemas/AtomicPromptTuple" }
        intent: { type: string }
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    HarnessF4Input:
      type: object
      required: [sessionId, sourceArtifactId, targetVibe]
      properties:
        sessionId: { type: string, format: uuid }
        sourceArtifactId: { type: string, format: uuid }
        targetVibe: { type: string }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    MicroPdd:
      type: object
      required: [cheatSheet, worksheet, buildLaunch, interfaceContract]
      properties:
        cheatSheet: { type: string }
        worksheet: { type: string }
        buildLaunch: { type: string }
        interfaceContract: { type: string }
        artifactId: { type: ["string", "null"], format: uuid }
    HarnessF5Input:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string, format: uuid }
        step: { type: integer, minimum: 1, maximum: 7 }
        answers:
          type: object
          additionalProperties: true
        finalize: { type: boolean }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    HarnessF5FinalizeInput:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string, format: uuid }
        name: { type: ["string", "null"], maxLength: 120 }
        answers:
          type: object
          additionalProperties: true
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    F5StepOutcome:
      type: object
      required: [done]
      properties:
        done: { type: boolean }
        nextStep: { type: ["integer", "null"] }
        nextQuestion: { type: ["string", "null"] }
        spc:
          oneOf:
            - { $ref: "#/components/schemas/Spc" }
            - { type: "null" }
    HarnessF6Input:
      type: object
      required: [sessionId, mode]
      properties:
        sessionId: { type: string, format: uuid }
        mode: { type: string, enum: [FROM_SPC, FRESH] }
        sourceArtifactId: { type: ["string", "null"], format: uuid }
        brief: { type: ["string", "null"] }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    HarnessF6FinalizeInput:
      type: object
      required: [sessionId, mode]
      properties:
        sessionId: { type: string, format: uuid }
        mode: { type: string, enum: [FROM_SPC, FRESH] }
        sourceArtifactId: { type: ["string", "null"], format: uuid }
        brief: { type: ["string", "null"] }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    AtlasPdd:
      type: object
      required: [cheatSheet, execSummary, worksheet, implementation]
      properties:
        cheatSheet: { type: string }
        execSummary: { type: string }
        worksheet: { type: string }
        implementation: { type: string }
        artifactId: { type: ["string", "null"], format: uuid }
    HarnessF6VdjInput:
      type: object
      required: [sessionId, pddArtifactId]
      properties:
        sessionId: { type: string, format: uuid }
        pddArtifactId: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/LlmProvider" }
    VdjRecommendation:
      type: object
      required: [recommendedIde, recommendedVibe, rationale]
      properties:
        recommendedIde: { type: string }
        recommendedVibe: { type: string }
        rationale: { type: string }
        alternatives:
          type: array
          items:
            type: object
            required: [name, fit]
            properties:
              name: { type: string }
              fit: { type: number }
    HarnessF7Input:
      type: object
      required: [sessionId, pddArtifactId]
      properties:
        sessionId: { type: string, format: uuid }
        pddArtifactId: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    HarnessF05Input:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string, format: uuid }
        brief: { type: ["string", "null"] }
        provider: { $ref: "#/components/schemas/LlmProvider" }
...
    MathmonIntake:
      type: object
      required: [id, sessionId, report, createdAt]
      properties:
        id: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        report: { $ref: "#/components/schemas/MathmonIntakeReport" }
        provider: { type: ["string", "null"] }
        modelId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    HarnessMapInput:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/LlmProvider" }
...